refactor: my, sign-up 폴더, 헤더 이름 정리
#21
